Adressing review remarks
